QA/QC area calcs
Percent protection was previous showing values >100% due using the area of the protected land computed before the clipping to the cluster/fa geometry. This resulting in the computation of percent protection based on all the land that *intersects* with the cluster/fa and not just what is within. These calculations may need to be updated if the geometry of FieldDoc Practice ID 51769 is updated.